Add hasLocationPermission check
